refactor: remove deeppavlov code
