--- annotators/ner/preproc_ner_rus.json
@@ -1,151 +1,0 @@
-{
-  "chainer": {
-    "in": ["last_utterances"],
-    "pipe": [
-      {
-        "in": ["last_utterances"],
-        "class_name": "lazy_tokenizer",
-        "out": ["tokens"]
-      },
-      {
-        "in": ["tokens"],
-        "class_name": "str_lower",
-        "out": ["x_lower"]
-      },
-      {
-        "in": ["x_lower"],
-        "class_name": "sanitizer",
-        "nums": true,
-        "out": ["x_san"]
-      },
-      {
-        "in": ["x_san"],
-        "id": "word_vocab",
-        "class_name": "simple_vocab",
-        "pad_with_zeros": true,
-        "special_tokens": ["<UNK>"],
-        "save_path": "{MODELS_PATH}/ner_rus/word.dict",
-        "load_path": "{MODELS_PATH}/ner_rus/word.dict",
-        "out": ["x_tok_ind"]
-      },
-      {
-        "id": "tag_vocab",
-        "class_name": "simple_vocab",
-        "pad_with_zeros": true,
-        "save_path": "{MODELS_PATH}/ner_rus/tag.dict",
-        "load_path": "{MODELS_PATH}/ner_rus/tag.dict",
-        "out": ["y_ind"]
-      },
-      {
-        "in": ["tokens"],
-        "class_name": "char_splitter",
-        "out": ["x_char"]
-      },
-      {
-        "in": ["x_char"],
-        "id": "char_vocab",
-        "class_name": "simple_vocab",
-        "pad_with_zeros": true,
-        "save_path": "{MODELS_PATH}/ner_rus/char.dict",
-        "load_path": "{MODELS_PATH}/ner_rus/char.dict",
-        "out": ["x_char_ind"]
-      },
-      {
-        "in": ["x_san"],
-        "id": "embedder",
-        "class_name": "fasttext",
-        "pad_zero": true,
-        "load_path": "{DOWNLOADS_PATH}/embeddings/lenta_lower_100.bin",
-        "out": ["x_emb"]
-      },
-      {
-        "in": ["tokens"],
-        "class_name": "mask",
-        "out": ["mask"]
-      },
-      {
-        "class_name": "random_emb_mat",
-        "id": "embeddings",
-        "vocab_len": "#word_vocab.len",
-        "emb_dim": 100
-      },
-      {
-        "class_name": "random_emb_mat",
-        "id": "embeddings_char",
-        "vocab_len": "#char_vocab.len",
-        "emb_dim": 100
-      },
-      {
-        "id": "capitalization",
-        "class_name": "capitalization_featurizer",
-        "in": ["tokens"],
-        "out": ["cap"]
-      },
-      {
-        "in": [
-          "x_emb",
-          "mask",
-          "x_char_ind",
-          "cap"
-        ],
-        "out": ["y_predicted"],
-        "class_name": "ner",
-        "main": true,
-        "n_hidden_list": [128],
-        "net_type": "rnn",
-        "cell_type": "lstm",
-        "use_cudnn_rnn": true,
-        "n_tags": "#tag_vocab.len",
-        "capitalization_dim": "#capitalization.dim",
-        "token_emb_dim": "#embedder.dim",
-        "char_emb_dim": 100,
-        "save_path": "{MODELS_PATH}/ner_rus/model",
-        "load_path": "{MODELS_PATH}/ner_rus/model",
-        "char_emb_mat": "#embeddings_char.emb_mat",
-        "use_crf": true,
-        "use_batch_norm": true,
-        "embeddings_dropout": true,
-        "top_dropout": true,
-        "intra_layer_dropout": true,
-        "l2_reg": 0,
-        "learning_rate": 1e-2,
-        "dropout_keep_prob": 0.7
-      },
-      {
-        "ref": "tag_vocab",
-        "in": ["y_predicted"],
-        "out": ["tags"]
-      }
-    ],
-    "out": [
-      "tokens",
-      "tags"
-    ]
-  },
-  "metadata": {
-    "variables": {
-      "ROOT_PATH": "~/.deeppavlov",
-      "DOWNLOADS_PATH": "{ROOT_PATH}/downloads",
-      "MODELS_PATH": "{ROOT_PATH}/models"
-    },
-    "requirements": [
-      "{DEEPPAVLOV_PATH}/requirements/fasttext.txt",
-      "{DEEPPAVLOV_PATH}/requirements/tf.txt"
-    ],
-    "labels": {
-      "telegram_utils": "NERCoNLL2003Model",
-      "server_utils": "NER"
-    },
-    "download": [
-      {
-        "url": "http://files.deeppavlov.ai/deeppavlov_data/ner_rus_v2_cpu_compatible.tar.gz",
-        "subdir": "{MODELS_PATH}"
-      },
-      {
-        "url": "http://files.deeppavlov.ai/embeddings/lenta_lower_100.bin",
-        "subdir": "{DOWNLOADS_PATH}/embeddings"
-      }
-    ]
-  }
-}
-
