--- annotators/obscenity/obscenity_classifier.json
@@ -1,30 +1,0 @@
-{
-  "chainer": {
-    "in": [
-      "last_utterances"
-    ],
-    "pipe": [
-      {
-        "class_name": "annotators.obscenity.obscenity_classifier:ObscenityClassifier",
-        "data_path": "{DOWNLOADS_PATH}/obscenity_dataset/",
-        "in": "last_utterances",
-        "out": "labels"
-      }
-    ],
-    "out": [
-      "labels"
-    ]
-  },
-  "metadata": {
-    "variables": {
-      "ROOT_PATH": "~/.deeppavlov",
-      "DOWNLOADS_PATH": "{ROOT_PATH}/downloads"
-    },
-    "download": [
-      {
-        "url": "http://files.deeppavlov.ai/models/obscenity_classifier/ru_obscenity_dataset.zip",
-        "subdir": "{DOWNLOADS_PATH}/obscenity_dataset"
-      }
-    ]
-  }
-}
