+++ core/agent.py
@@ -5,7 +5,7 @@ import operator
 
 from core.state_manager import StateManager, get_state
 from core.skill_manager import SkillManager
-from core.hardcode_utterances import TG_START_UTT
+from models.hardcode_utterances import TG_START_UTT
 from core.state_schema import Dialog, Human
 
 Profile = Dict[str, Any]
