+++ dp/dockerfile_skill_basic
@@ -19,6 +19,6 @@ ENV PYTHONPATH "${PYTONPATH}:/dp-agent"
 
 RUN if [ $GPU = "true" ]; then pip install -r /base/DeepPavlov/deeppavlov/requirements/tf-gpu.txt; fi
 RUN python3.6 -m deeppavlov install $CONFIG
-RUN python3.6 dp_server_config.py
+RUN python3.6 dp/dp_server_config.py
 
 ENTRYPOINT python3.6 -m deeppavlov riseapi $CONFIG -p $PORT -d
