--- skill_selectors/chitchat_odqa_selector/sselector_chitchat_odqa.json
@@ -1,180 +1,0 @@
-{
-  "chainer": {
-    "in": [
-      "last_utterances"
-    ],
-    "in_y": [
-      "y"
-    ],
-    "pipe": [
-      {
-        "id": "classes_vocab",
-        "class_name": "simple_vocab",
-        "fit_on": [
-          "y"
-        ],
-        "save_path": "{MODELS_PATH}/classifiers/chitchat_vs_odqa_v5/classes.dict",
-        "load_path": "{MODELS_PATH}/classifiers/chitchat_vs_odqa_v5/classes.dict",
-        "in": "y",
-        "out": "y_ids"
-      },
-      {
-        "in": [
-          "last_utterances"
-        ],
-        "out": [
-          "x_prep"
-        ],
-        "class_name": "dirty_comments_preprocessor",
-        "remove_punctuation": true
-      },
-      {
-        "in": "x_prep",
-        "out": "x_tok",
-        "id": "my_tokenizer",
-        "class_name": "nltk_moses_tokenizer"
-      },
-      {
-        "in": "x_tok",
-        "out": "x_emb",
-        "id": "my_embedder",
-        "class_name": "fasttext",
-        "load_path": "{DOWNLOADS_PATH}/embeddings/ft_native_300_ru_wiki_lenta_nltk_wordpunct_tokenize.bin",
-        "dim": 300,
-        "pad_zero": true
-      },
-      {
-        "in": "y_ids",
-        "out": "y_onehot",
-        "class_name": "one_hotter",
-        "id": "my_one_hotter",
-        "depth": "#classes_vocab.len"
-      },
-      {
-        "in": [
-          "x_emb"
-        ],
-        "in_y": [
-          "y_onehot"
-        ],
-        "out": [
-          "y_pred_probas"
-        ],
-        "main": true,
-        "class_name": "keras_classification_model",
-        "save_path": "{MODELS_PATH}/classifiers/chitchat_vs_odqa_v5/model",
-        "load_path": "{MODELS_PATH}/classifiers/chitchat_vs_odqa_v5/model",
-        "embedding_size": "#my_embedder.dim",
-        "n_classes": "#classes_vocab.len",
-        "units_gru": 256,
-        "optimizer": "Adam",
-        "learning_rate": 0.01,
-        "learning_rate_decay": 0.01,
-        "loss": "categorical_crossentropy",
-        "coef_reg_gru": 1e-6,
-        "coef_reg_den": 1e-6,
-        "dropout_rate": 0.2,
-        "rec_dropout_rate": 0.2,
-        "dense_size": 100,
-        "model_name": "bigru_with_max_aver_pool_model",
-        "last_layer_activation": "softmax",
-        "restore_lr": false
-      },
-      {
-        "in": "y_pred_probas",
-        "out": "y_pred_ids",
-        "class_name": "proba2labels",
-        "max_proba": true
-      },
-      {
-        "in": "y_pred_ids",
-        "out": "y_pred_labels",
-        "ref": "classes_vocab"
-      },
-      {
-        "ref": "my_one_hotter",
-        "in": "y_pred_ids",
-        "out": "y_pred_onehot"
-      },
-      {
-        "class_name": "skill_selectors.chitchat_odqa_selector.keras_probas_converter:KerasClassificationProbasConverter",
-        "in": "y_pred_probas",
-        "out": "skill_names"
-      }
-    ],
-    "out": [
-      "skill_names"
-    ]
-  },
-  "train": {
-    "epochs": 10,
-    "batch_size": 32,
-    "metrics": [
-      {
-        "name": "roc_auc",
-        "inputs": [
-          "y_onehot",
-          "y_pred_probas"
-        ]
-      },
-      {
-        "name": "f1_macro",
-        "inputs": [
-          "y",
-          "y_pred_labels"
-        ]
-      },
-      {
-        "name": "f1_weighted",
-        "inputs": [
-          "y",
-          "y_pred_labels"
-        ]
-      },
-      {
-        "name": "sets_accuracy",
-        "inputs": [
-          "y",
-          "y_pred_labels"
-        ]
-      }
-    ],
-    "validation_patience": 5,
-    "val_every_n_epochs": 1,
-    "log_every_n_epochs": 1,
-    "show_examples": false,
-    "validate_best": true,
-    "test_best": true,
-    "tensorboard_log_dir": "{MODELS_PATH}"
-  },
-  "metadata": {
-    "variables": {
-      "ROOT_PATH": "~/.deeppavlov",
-      "DOWNLOADS_PATH": "{ROOT_PATH}/downloads",
-      "MODELS_PATH": "{ROOT_PATH}/models"
-    },
-    "requirements": [
-      "{DEEPPAVLOV_PATH}/requirements/tf.txt",
-      "{DEEPPAVLOV_PATH}/requirements/tf-hub.txt",
-      "{DEEPPAVLOV_PATH}/requirements/fasttext.txt"
-    ],
-    "labels": {
-      "telegram_utils": "IntentModel",
-      "server_utils": "KerasIntentModel"
-    },
-    "download": [
-      {
-        "url": "http://files.deeppavlov.ai/deeppavlov_data/classifiers/chitchat_vs_odqa_v5.tar.gz",
-        "subdir": "{ROOT_PATH}/models/classifiers/"
-      },
-      {
-        "url": "http://files.deeppavlov.ai/embeddings/ft_native_300_ru_wiki_lenta_nltk_wordpunct_tokenize/ft_native_300_ru_wiki_lenta_nltk_wordpunct_tokenize.bin",
-        "subdir": "{DOWNLOADS_PATH}/embeddings"
-      },
-      {
-        "url": "http://files.deeppavlov.ai/datasets/chitchat_vs_odqa.tar.gz",
-        "subdir": "{DOWNLOADS_PATH}/"
-      }
-    ]
-  }
-}
