--- skills/autofaq/agent_tfidf_autofaq.json
@@ -1,97 +1,0 @@
-{
-  "dataset_reader": {
-    "class_name": "faq_reader",
-    "x_col_name": "Question",
-    "y_col_name": "Answer",
-    "data_url": "http://files.deeppavlov.ai/faq/school/faq_school.csv"
-  },
-  "dataset_iterator": {
-    "class_name": "data_learning_iterator"
-  },
-  "chainer": {
-    "in": "dialogs",
-    "pipe": [
-      {
-        "class_name": "core.state_parsers.state_parser:DialogsParser",
-        "in": ["dialogs"],
-        "out": [
-          "last_utterances",
-          "last_annotations",
-          "utterances_histories",
-          "annotations_histories",
-          "dialog_ids",
-          "user_ids"
-        ]
-      },
-      {
-        "class_name": "ru_tokenizer",
-        "id": "my_tokenizer",
-        "in": "last_utterances",
-        "lemmas": true,
-        "out": "q_token_lemmas"
-      },
-      {
-        "ref": "my_tokenizer",
-        "in": "q_token_lemmas",
-        "out": "q_lem"
-      },
-      {
-        "in": [
-          "q_lem"
-        ],
-        "out": [
-          "q_vect"
-        ],
-        "fit_on": [
-          "q_lem"
-        ],
-        "id": "tfidf_vec",
-        "class_name": "sklearn_component",
-        "save_path": "{MODELS_PATH}/vectorizer/tfidf_vectorizer_ruwiki.pkl",
-        "load_path": "{MODELS_PATH}/vectorizer/tfidf_vectorizer_ruwiki.pkl",
-        "model_class": "sklearn.feature_extraction.text:TfidfVectorizer",
-        "infer_method": "transform"
-      },
-      {
-        "class_name": "cos_sim_classifier",
-        "in": "q_vect",
-        "fit_on": [
-          "q_vect",
-          "y"
-        ],
-        "top_n": 1,
-        "save_path": "{MODELS_PATH}/faq/tfidf_cos_sim_classifier.pkl",
-        "load_path": "{MODELS_PATH}/faq/tfidf_cos_sim_classifier.pkl",
-        "out": [
-          "text",
-          "confidence"
-        ]
-      }
-    ],
-    "out": [
-      "text",
-      "confidence"
-    ]
-  },
-  "train": {
-    "validate_best": false,
-    "test_best": false
-  },
-  "metadata": {
-    "variables": {
-      "ROOT_PATH": "~/.deeppavlov",
-      "DOWNLOADS_PATH": "{ROOT_PATH}/downloads",
-      "MODELS_PATH": "{ROOT_PATH}/models"
-    },
-    "download": [
-      {
-        "url": "http://files.deeppavlov.ai/faq/school/tfidf_cos_sim_classifier.pkl",
-        "subdir": "{MODELS_PATH}/faq"
-      },
-      {
-        "url": "http://files.deeppavlov.ai/vectorizer/tfidf_vectorizer_ruwiki.pkl",
-        "subdir": "{MODELS_PATH}/vectorizer"
-      }
-    ]
-  }
-}
