--- skills/ranking_chitchat/agent_ranking_chitchat_2staged_tfidf_smn_v4_prep.json
@@ -1,139 +1,0 @@
-{
-  "info": "(v4) Agent_v2-based ranking chit-chat. Using cleaned dataset Books + VK 2-turns dialogs + pBot. There are 279901 items in total.",
-  "chainer": {
-    "in": ["utterances_batch",
-           "last_annotations",
-           "history_batch",
-           "dialogs"],
-    "pipe": [
-      {
-        "class_name": "skills.ranking_chitchat.parsers:AnnotationsParser",
-        "keys": ["ner.tokens", "ner.tags"],
-        "in": ["last_annotations"],
-        "out": ["utterance_toks", "utterance_tags"]
-      },
-      {
-        "info": "0. Detect if user introduced himself",
-        "class_name": "annotators.person.person_normalizer:MyselfDetector",
-        "person_tag": "PER",
-        "in": ["utterance_toks", "utterance_tags", "dialogs"],
-        "out": ["name"]
-      },
-      {
-        "info": "1. Replace separators from input string",
-        "class_name": "skills.ranking_chitchat.compose_inputs_hybrid_ranker:ComposeInputsHybridRanker",
-        "context_depth": 3,
-        "history_includes_last_utterance": true,
-        "use_user_context_only": true,
-        "use_context_for_query": false,
-        "in": ["utterances_batch", "history_batch"],
-        "out": ["query_batch", "expanded_context_batch"]
-      },
-      {
-        "class_name": "hashing_tfidf_vectorizer",
-        "id": "vectorizer",
-        "save_path": "{MODELS_PATH}/chitchat_2staged_tfidf_smn_v3/train_contexts_tfidf_matrix.npz",
-        "load_path": "{MODELS_PATH}/chitchat_2staged_tfidf_smn_v3/train_contexts_tfidf_matrix.npz",
-        "tokenizer": {"class_name": "ru_tokenizer",
-         "alphas_only": false,
-         "lemmas": true,
-         "ngram_range": [1, 2, 3, 4, 5]}
-      },
-      {
-        "info": "2. Find relevant contexts",
-         "class_name": "tfidf_ranker",
-        "top_n": 500,
-        "in": ["query_batch"],
-        "out": ["tfidf_doc_ids", "tfidf_doc_scores"],
-        "vectorizer": "#vectorizer"
-      },
-      {
-        "info": "3. Retrieve response candidates using the context retrieved",
-        "class_name": "skills.ranking_chitchat.retrieve_candidates:RetrieveCandidates",
-        "map_filename": "{MODELS_PATH}/chitchat_2staged_tfidf_smn_v3/train_index_dict_v3.pickle",
-        "in": ["expanded_context_batch", "tfidf_doc_ids", "tfidf_doc_scores"],
-        "out": ["response_candidates", "model_inputs"]
-      },
-      {
-        "info": "4. Preprocess response candidates and choose the best one",
-        "id": "preproc",
-        "class_name": "siamese_preprocessor",
-        "load_path": "{MODELS_PATH}/chitchat_2staged_tfidf_smn_v3/preproc/tok.dict",
-        "num_ranking_samples": 500,
-        "num_context_turns": 10,
-        "max_sequence_length": 50,
-        "embedding_dim": 200,
-        "in": ["model_inputs"],
-        "out": ["x_proc"],
-        "tokenizer": {
-          "class_name": "ru_tokenizer",
-          "id": "tok_1"
-        },
-        "vocab": {
-          "class_name": "simple_vocab",
-          "special_tokens": ["", "<UNK>"],
-          "unk_token": "<UNK>",
-          "id": "vocab_1",
-          "load_path": "{MODELS_PATH}/chitchat_2staged_tfidf_smn_v3/vocabs/int_tok.dict",
-          "save_path": "{MODELS_PATH}/chitchat_2staged_tfidf_smn_v3/vocabs/int_tok.dict"
-        },
-        "embedder": {
-          "id": "word2vec_embedder",
-          "class_name": "glove",
-          "dim": 200,
-          "load_path": "{MODELS_PATH}/chitchat_2staged_tfidf_smn_v3/chitchat1.w2v.vectors.txt"
-        }
-      },
-      {
-        "id": "embeddings",
-        "class_name": "emb_mat_assembler",
-        "embedder": "#word2vec_embedder",
-        "vocab": "#vocab_1"
-      },
-      {
-        "in": ["x_proc"],
-        "out": ["y_predicted"],
-        "class_name": "smn_nn",
-        "num_context_turns": "#preproc.num_context_turns",
-        "max_sequence_length": "#preproc.max_sequence_length",
-        "embedding_dim": "#word2vec_embedder.dim",
-        "emb_matrix": "#embeddings.emb_mat",
-        "learning_rate": 1e-3,
-        "batch_size": 500,
-        "seed": 65,
-        "load_path": "{MODELS_PATH}/chitchat_2staged_tfidf_smn_v3/model_smn/model",
-        "save_path": "{MODELS_PATH}/chitchat_2staged_tfidf_smn_v3/model_smn/model"
-      },
-      {
-        "info": "5. Return actual response",
-        "class_name": "skills.ranking_chitchat.hybrid_ranker_predictor:HybridRankerPredictor",
-        "sample_size": 5,
-        "in": ["response_candidates", "y_predicted"],
-        "out": ["text", "confidence"]
-      }
-    ],
-    "out": ["text", "confidence", "name"]
-  },
-  "metadata": {
-    "variables": {
-      "ROOT_PATH": "~/.deeppavlov",
-      "DOWNLOADS_PATH": "{ROOT_PATH}/downloads",
-      "MODELS_PATH": "{ROOT_PATH}/models",
-      "CONFIGS_PATH": "{DEEPPAVLOV_PATH}/configs"
-    },
-    "requirements": [
-      "{DEEPPAVLOV_PATH}/requirements/tf.txt",
-      "{DEEPPAVLOV_PATH}/requirements/fasttext.txt",
-      "{DEEPPAVLOV_PATH}/requirements/gensim.txt"
-    ],
-    "labels": {
-      "server_utils": "Chitchat"
-    },
-    "download": [
-      {
-        "url": "http://files.deeppavlov.ai/deeppavlov_data/dev/chitchat_2staged_tfidf_smn_v3.tar.gz",
-        "subdir": "{MODELS_PATH}"
-      }
-    ]
-  }
-}
