--- skills/text_qa/agent_ranking_mailru_bert_3.json
@@ -1,99 +1,0 @@
-{
-  "chainer": {
-    "in": ["dialogs"],
-    "out": ["text", "confidence"],
-
-    "pipe": [
-      {
-        "class_name": "core.state_parsers.state_parser:DialogsParser",
-        "in": ["dialogs"],
-        "out": [
-          "last_utterances",
-          "last_annotations",
-          "utterances_histories",
-          "annotations_histories",
-          "dialog_ids",
-          "user_ids"
-        ]
-      },
-      {
-        "class_name": "response_base_loader",
-        "id": "loader",
-        "save_path": "{DOWNLOADS_PATH}/mailru_bert_npair_simple_norm_vec",
-        "load_path": "{DOWNLOADS_PATH}/mailru_bert_npair_simple_norm_vec"
-      },
-      {
-        "class_name": "skills.text_qa.bert_ranker_preprocessor:BertRankerPreprocessor",
-        "id": "preproc",
-        "vocab_file": "{DOWNLOADS_PATH}/mailru_bert_npair_simple_norm_vec/vocab.txt",
-        "do_lower_case": true,
-        "max_seq_length": 128,
-        "resps": "#loader.resps",
-        "resp_vecs": "#loader.resp_vecs",
-        "conts": "#loader.conts",
-        "cont_vecs": "#loader.cont_vecs",
-        "in": ["last_utterances"],
-        "out": ["bert_features"]
-      },
-      {
-        "class_name": "skills.text_qa.bert_ranker:BertRankerModel",
-        "resps": "#loader.resps",
-        "resp_vecs": "#loader.resp_vecs",
-        "resp_features": "#preproc.resp_features",
-        "conts": "#loader.conts",
-        "cont_vecs": "#loader.cont_vecs",
-        "cont_features": "#preproc.cont_features",
-        "bot_mode": 2,
-        "n_classes": 2,
-        "one_hot_labels": false,
-        "bert_config_file": "{DOWNLOADS_PATH}/mailru_bert_npair_simple_norm_vec/bert_config.json",
-        "pretrained_bert": "{DOWNLOADS_PATH}/mailru_bert_npair_simple_norm_vec/model.ckpt-599058",
-        "save_path": "{DOWNLOADS_PATH}/mailru_bert_npair_simple_norm_vec",
-        "load_path": "{DOWNLOADS_PATH}/mailru_bert_npair_simple_norm_vec/model.ckpt-599058",
-        "num_ranking_samples": 1,
-        "batch_size": 32,
-        "keep_prob": 1.0,
-        "attention_probs_keep_prob": 1.0,
-        "hidden_keep_prob": 1.0,
-        "learning_rate": 2e-05,
-        "learning_rate_drop_patience": 2,
-        "learning_rate_drop_div": 2.0,
-        "in": ["bert_features"],
-        "out": ["text", "confidence"]
-      }
-    ]
-  },
-  "train": {
-    "batch_size": 32,
-    "pytest_max_batches": 2,
-    "train_metrics": [],
-    "metrics": ["r@1", "r@2", "r@5"],
-    "validation_patience": 10,
-    "val_every_n_batches": 100,
-    "log_every_n_batches": 100,
-    "validate_best": true,
-    "test_best": true,
-    "tensorboard_log_dir": "{MODELS_PATH}/mailru_bert_model/logs"
-  },
-  "metadata": {
-    "variables": {
-      "ROOT_PATH": "~/.deeppavlov",
-      "DOWNLOADS_PATH": "{ROOT_PATH}/downloads",
-      "MODELS_PATH": "{ROOT_PATH}/models"
-    },
-    "requirements": [
-      "{DEEPPAVLOV_PATH}/requirements/tf.txt",
-      "{DEEPPAVLOV_PATH}/requirements/bert_dp.txt"
-    ],
-    "download": [
-      {
-        "url": "http://files.deeppavlov.ai/datasets/mailru_bert.tar.gz",
-        "subdir": "{DOWNLOADS_PATH}"
-      },
-      {
-        "url": "http://files.deeppavlov.ai/deeppavlov_data/mailru_bert_npair_simple_norm_vec.tar.gz",
-        "subdir": "{DOWNLOADS_PATH}/"
-      }
-    ]
-  }
-}
