--- skills/text_qa/agent_ru_odqa_retr_noans_rubert_infer.json
@@ -1,78 +1,0 @@
-{
-  "chainer": {
-    "in": [
-      "questions"
-    ],
-    "out": [
-      "text",
-      "confidence"
-    ],
-    "pipe": [
-      {
-        "config_path": "{CONFIGS_PATH}/doc_retrieval/ru_ranker_tfidf_wiki.json",
-        "in": [
-          "questions"
-        ],
-        "out": [
-          "tfidf_doc_ids"
-        ]
-      },
-      {
-        "class_name": "wiki_sqlite_vocab",
-        "in": [
-          "tfidf_doc_ids"
-        ],
-        "out": [
-          "tfidf_doc_text"
-        ],
-        "join_docs": false,
-        "shuffle": false,
-        "load_path": "{DOWNLOADS_PATH}/odqa/ruwiki.db"
-      },
-      {
-        "class_name": "document_chunker",
-        "in": ["tfidf_doc_text"],
-        "out": ["chunks"],
-        "flatten_result": true,
-        "paragraphs": true
-      },
-      {
-        "class_name": "string_multiplier",
-        "in": ["questions", "chunks"],
-        "out":["questions"]
-      },
-      {
-        "class_name": "logit_ranker",
-        "batch_size": 10,
-        "squad_model": {"config_path": "{CONFIGS_PATH}/squad/multi_squad_ru_retr_noans_rubert_infer.json"},
-        "sort_noans": true,
-        "norm_coef": 10000.0,
-        "in": [
-          "chunks",
-          "questions"
-        ],
-        "out": [
-          "text",
-          "confidence"
-        ]
-      }
-    ]
-  },
-  "metadata": {
-    "variables": {
-      "ROOT_PATH": "~/.deeppavlov",
-      "DOWNLOADS_PATH": "{ROOT_PATH}/downloads",
-      "MODELS_PATH": "{ROOT_PATH}/models",
-      "CONFIGS_PATH": "{DEEPPAVLOV_PATH}/configs"
-    },
-    "requirements": [
-      "{DEEPPAVLOV_PATH}/requirements/tf.txt",
-      "{DEEPPAVLOV_PATH}/requirements/bert_dp.txt"
-    ],
-    "labels": {
-      "server_utils": "ODQA"
-    },
-    "download": [
-    ]
-  }
-}
