+++ tests/agent_test.py
@@ -5,8 +5,8 @@ from core.agent import Agent
 from core.state_manager import StateManager
 from core.skill_manager import SkillManager
 from core.rest_caller import RestCaller
-from core.postprocessor import DefaultPostprocessor
-from core.response_selector import ConfidenceResponseSelector
+from models.postprocessor import DefaultPostprocessor
+from models.response_selector import ConfidenceResponseSelector
 from core.transform_config import MAX_WORKERS, ANNOTATORS, SKILL_SELECTORS
 from core.skill_selector import ChitchatQASelector
 from core.state_schema import Human
